+++ backend/src/apiserver/server/util.go
@@ -7,14 +7,20 @@ import (
 	"bytes"
 	"compress/gzip"
 	"encoding/json"
-	"io"
-	"io/ioutil"
-	"net/url"
-	"strings"
-
+	"github.com/argoproj/argo/errors"
+	"github.com/argoproj/argo/pkg/apis/workflow"
+	wfv1 "github.com/argoproj/argo/pkg/apis/workflow/v1alpha1"
 	api "github.com/kubeflow/pipelines/backend/api/go_client"
 	"github.com/kubeflow/pipelines/backend/src/apiserver/resource"
 	"github.com/kubeflow/pipelines/backend/src/common/util"
+	"io"
+	"io/ioutil"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"math/rand"
+	"net/url"
+	"regexp"
+	"strings"
+	"time"
 )
 
 // These are valid conditions of a ScheduledWorkflow.
@@ -22,6 +28,8 @@ const (
 	MaxFileNameLength = 100
 	MaxFileLength     = 32 << 20 // 32Mb
 )
+
+const letters = "abcdefghijklmnopqrstuvwxyz0123456789"
 
 // This method extract the common logic of naming the pipeline.
 // API caller can either explicitly name the pipeline through query string ?name=foobar
@@ -240,4 +248,92 @@ func ValidatePipelineSpec(resourceManager *resource.ResourceManager, spec *api.P
 		return util.NewInvalidInputError("The input parameter length exceed maximum size of %v.", util.MaxParameterBytes)
 	}
 	return nil
+}
+
+func randString(n int) string {
+	b := make([]byte, n)
+	for i := range b {
+		b[i] = letters[rand.Intn(len(letters))]
+	}
+	return string(b)
+}
+
+// convertNodeID converts an old nodeID to a new nodeID
+func convertNodeID(newWf *wfv1.Workflow, regex *regexp.Regexp, oldNodeID string, oldNodes map[string]wfv1.NodeStatus) string {
+	node := oldNodes[oldNodeID]
+	newNodeName := regex.ReplaceAllString(node.Name, newWf.ObjectMeta.Name)
+	return newWf.NodeID(newNodeName)
+}
+
+// FormulateResubmitWorkflow formulate a new workflow from a previous workflow optionally re-using successful nodes
+func formulateResubmitWorkflow(wf *wfv1.Workflow) (*wfv1.Workflow, error) {
+	newWF := wfv1.Workflow{}
+	newWF.TypeMeta = wf.TypeMeta
+
+	// Resubmitted workflow will use generated names
+	if wf.ObjectMeta.GenerateName != "" {
+		newWF.ObjectMeta.GenerateName = wf.ObjectMeta.GenerateName
+	} else {
+		newWF.ObjectMeta.GenerateName = wf.ObjectMeta.Name + "-"
+	}
+	// When resubmitting workflow with memoized nodes, we need to use a predetermined workflow name
+	// in order to formulate the node statuses. Which means we cannot reuse metadata.generateName
+	// The following simulates the behavior of generateName
+	newWF.ObjectMeta.Name = newWF.ObjectMeta.GenerateName + randString(5)
+
+	// carry over the unmodified spec
+	newWF.Spec = wf.Spec
+
+	// carry over user labels and annotations from previous workflow.
+	// skip any argoproj.io labels except for the controller instanceID label.
+	for key, val := range wf.ObjectMeta.Labels {
+		if strings.HasPrefix(key, workflow.FullName+"/") && key != workflow.FullName+"/controller-instanceid" {
+			continue
+		}
+		if newWF.ObjectMeta.Labels == nil {
+			newWF.ObjectMeta.Labels = make(map[string]string)
+		}
+		newWF.ObjectMeta.Labels[key] = val
+	}
+	for key, val := range wf.ObjectMeta.Annotations {
+		if newWF.ObjectMeta.Annotations == nil {
+			newWF.ObjectMeta.Annotations = make(map[string]string)
+		}
+		newWF.ObjectMeta.Annotations[key] = val
+	}
+
+	// Iterate the previous nodes. If it was successful Pod carry it forward
+	replaceRegexp := regexp.MustCompile("^" + wf.ObjectMeta.Name)
+	newWF.Status.Nodes = make(map[string]wfv1.NodeStatus)
+	for _, node := range wf.Status.Nodes {
+		switch node.Phase {
+		case wfv1.NodeSucceeded, wfv1.NodeSkipped:
+			node.Name = replaceRegexp.ReplaceAllString(node.Name, newWF.ObjectMeta.Name)
+			node.BoundaryID = convertNodeID(&newWF, replaceRegexp, node.BoundaryID, wf.Status.Nodes)
+			node.StartedAt = metav1.Time{Time: time.Now().UTC()}
+			node.FinishedAt = node.StartedAt
+			newChildren := make([]string, len(node.Children))
+			for i, childID := range node.Children {
+				newChildren[i] = convertNodeID(&newWF, replaceRegexp, childID, wf.Status.Nodes)
+			}
+			node.Children = newChildren
+			newOutboundNodes := make([]string, len(node.OutboundNodes))
+			for i, outboundID := range node.OutboundNodes {
+				newOutboundNodes[i] = convertNodeID(&newWF, replaceRegexp, outboundID, wf.Status.Nodes)
+			}
+			node.OutboundNodes = newOutboundNodes
+			if node.Type == wfv1.NodeTypePod {
+				node.Phase = wfv1.NodeSkipped
+				node.Type = wfv1.NodeTypeSkipped
+			}
+			newWF.Status.Nodes[node.ID] = node
+		case wfv1.NodeError, wfv1.NodeFailed, wfv1.NodeRunning:
+			// do not add this status to the node. pretend as if this node never existed.
+			// NOTE: NodeRunning shouldn't really happen except in weird scenarios where controller
+			// mismanages state (e.g. panic when operating on a workflow)
+		default:
+			return nil, errors.InternalErrorf("Workflow cannot be resubmitted with nodes in %s phase", node, node.Phase)
+		}
+	}
+	return &newWF, nil
 }
